# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/addresses": {
      "get": {
        "operationId": "AddressesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Address"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      },
      "post": {
        "operationId": "AddressesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Address"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
    "/api/v1/addresses/{id}": {
      "get": {
        "operationId": "AddressesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Address"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      },
      "patch": {
        "operationId": "AddressesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Address"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      },
      "delete": {
        "operationId": "AddressesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
    "/api/v1/addresses/{id}/default": {
      "patch": {
        "operationId": "AddressesController_setDefault_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
    "/api/v1/shop": {
      "post": {
        "operationId": "ShopController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Shop"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Shop"
        ]
      },
      "patch": {
        "operationId": "ShopController_update_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShopDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Shop"
        ]
      }
    },
...
    "/api/v1/shop/{sellerId}": {
      "get": {
        "operationId": "ShopController_getShop_v1",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Shop"
        ]
      }
    },
    "/api/v1/shop/{sellerId}/products": {
      "get": {
        "operationId": "ShopController_getProducts_v1",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Product"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Shop"
        ]
      }
    },
    "/api/v1/shop/{sellerId}/orders": {
      "get": {
        "operationId": "ShopController_getSellerOrders_v1",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Shop"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zoldify API",
    "description": "API của sàn mua bán đồ cũ Zoldify. Web (Next.js) và app (React Native) dùng chung hợp đồng này.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "current": {
            "type": "number",
            "description": "Trang hiện tại, bắt đầu từ 1"
          },
          "pageSize": {
            "type": "number",
            "description": "Số bản ghi mỗi trang"
          },
          "pages": {
            "type": "number",
            "description": "Tổng số trang"
          },
          "total": {
            "type": "number",
            "description": "Tổng số bản ghi khớp điều kiện lọc"
          }
        },
        "required": [
          "current",
          "pageSize",
          "pages",
          "total"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "role": {
            "enum": [
              "buyer",
              "seller",
              "admin",
              "moderator"
            ],
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "is_locked": {
            "type": "boolean"
          },
          "last_seen": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "role",
          "avatar",
          "email_verified",
          "is_locked",
          "last_seen",
          "gender",
          "created_at"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        },
        "required": [
          "id",
          "name",
          "name_en",
          "description",
          "slug",
          "image",
          "is_active",
          "created_at",
          "updated_at",
          "products"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "description": "4.1. Đơn vị tiền của `price` — mã ISO 4217 ('VND', 'USD', 'JPY'…).\n\nVÌ SAO CẦN: trước cột này, mọi con số trong hệ thống đều NGẦM hiểu là đồng\nViệt Nam, và không chỗ nào trong mã nói ra điều đó. Giao diện tiếng Anh vì\nvậy vẫn in \"1.890.000 ₫\" — không phải lỗi định dạng, mà là vì thật sự không\ncó dữ liệu nào cho nó biết số đó là tiền gì.\n\n⚠️ CỘT NÀY CHỈ LÀM ĐÚNG PHẦN HIỂN THỊ. Nó KHÔNG khiến sàn giao dịch được\nxuyên tiền tệ. Muốn thế còn cần: nguồn tỉ giá, thời điểm chốt giá, và một\nsổ cái nhiều tiền tệ. Sổ cái hiện tại (ledger_accounts) kiểm bất biến\n`sum(entries) = 0`, mà bất biến đó chỉ có nghĩa TRONG CÙNG một tiền tệ —\ntrộn hai loại tiền vào một sổ là cách làm hỏng sổ sách mà không ai thấy.\nNên tới khi có thiết kế đó, mỗi sàn vẫn nên chạy một tiền tệ."
          },
          "stock": {
            "type": "number"
          },
          "image": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "spec": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "condition": {
            "type": "string"
          },
          "is_freeship": {
            "type": "boolean"
          },
          "sold_count": {
            "type": "number"
          },
          "view_count": {
            "type": "number"
          },
          "status": {
            "enum": [
              "draft",
              "pending",
              "active",
              "sold",
              "rejected"
            ],
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "price",
          "currency",
          "stock",
          "image",
          "brand",
          "spec",
          "images",
          "condition",
          "is_freeship",
          "sold_count",
          "view_count",
          "status",
          "category",
          "seller",
          "created_at",
          "updated_at"
        ]
      },
...
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "recipient_name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "description": "Mã quốc gia ISO 3166-1 alpha-2 ('VN', 'US', 'JP'…).\n\nĐứng TRƯỚC province vì nó quyết định mấy ô còn lại có nghĩa gì: 'province /\ndistrict / ward' là ba cấp hành chính của Việt Nam. Nước khác chia khác —\nMỹ có state/city/ZIP, Nhật có prefecture/city/chōme. Cột này chưa làm cho\nbiểu mẫu tự đổi theo nước, nhưng nó ghi lại được ĐỊA CHỈ NÀY THUỘC NƯỚC NÀO,\nlà thứ tối thiểu phải có trước khi làm việc đó.\n\nMặc định 'VN' cho mọi dòng đang có: tất cả địa chỉ trong database lúc này\nđều nhập bằng bộ chọn tỉnh/huyện/xã Việt Nam, nên đó là sự thật chứ không\nphải phỏng đoán."
          },
          "province": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "recipient_name",
          "phone_number",
          "label",
          "country",
          "province",
          "district",
          "ward",
          "street",
          "is_default",
          "created_at",
          "updated_at"
        ]
      },
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "recipient_name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          }
        },
        "required": [
          "recipient_name",
          "phone_number",
          "province",
          "district",
          "street"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "recipient_name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          }
        }
      },
      "CreateShopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "banner": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "pickup_name": {
            "type": "string"
          },
          "pickup_phone": {
            "type": "string"
          },
          "pickup_address": {
            "type": "string"
          },
          "pickup_province_name": {
            "type": "string"
          },
          "pickup_district_id": {
            "type": "number"
          },
          "pickup_district_name": {
            "type": "string"
          },
          "pickup_ward_code": {
            "type": "string"
          },
          "pickup_ward_name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "Shop": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "banner": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "pickup_name": {
            "type": "string",
            "description": "Địa chỉ LẤY HÀNG của người bán — nguồn gửi cho vận đơn GHN.\n\nZoldify là sàn C2C: mỗi người bán tự gửi hàng từ nhà mình, KHÔNG có kho\nchung của sàn. Nên \"from\" của mỗi vận đơn phải là địa chỉ này, không phải\nmột điểm gửi cố định trong biến môi trường.\n\nLưu cả ID lẫn TÊN vì GHN dùng hai kiểu ở hai chỗ:\n  · tính phí  -> cần district_id (số) + ward_code\n  · tạo đơn   -> cần from_district_name / from_ward_name / from_province_name (chữ)\nCả hai lấy được cùng lúc từ danh mục /ghn/provinces|districts|wards, nên\nchụp lại luôn để khỏi tra ngược tên<->id về sau.\n\nNullable vì shop cũ chưa khai; chưa đủ pickup thì rơi về shop GHN mặc định\ncủa sàn (GHN_SHOP_ID) — xem GhnService. Bắt buộc khai đủ trước khi bán."
          },
          "pickup_phone": {
            "type": "string"
          },
          "pickup_address": {
            "type": "string"
          },
          "pickup_province_name": {
            "type": "string"
          },
          "pickup_district_id": {
            "type": "number"
          },
          "pickup_district_name": {
            "type": "string"
          },
          "pickup_ward_code": {
            "type": "string"
          },
          "pickup_ward_name": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "inactive",
              "banned"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "name",
          "slug",
          "description",
          "logo",
          "banner",
          "phone",
          "address",
          "pickup_name",
          "pickup_phone",
          "pickup_address",
          "pickup_province_name",
          "pickup_district_id",
          "pickup_district_name",
          "pickup_ward_code",
          "pickup_ward_name",
          "status",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateShopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "banner": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "pickup_name": {
            "type": "string"
          },
          "pickup_phone": {
            "type": "string"
          },
          "pickup_address": {
            "type": "string"
          },
          "pickup_province_name": {
            "type": "string"
          },
          "pickup_district_id": {
            "type": "number"
          },
          "pickup_district_name": {
            "type": "string"
          },
          "pickup_ward_code": {
            "type": "string"
          },
          "pickup_ward_name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}